requirements.txt
requests
pyyaml
bs4
pandas
argparse
boto3
datetime
lxml
shutil